Setup unit tests with react-testing-library
