+++ package.json
@@ -46,6 +46,8 @@
         "@babel/plugin-proposal-class-properties": "^7.8.3",
         "@babel/preset-env": "^7.8.4",
         "@babel/preset-react": "^7.8.3",
+        "@testing-library/jest-dom": "^5.11.4",
+        "@testing-library/react": "^11.0.4",
         "babel-eslint": "^10.1.0",
         "babel-loader": "^8.0.6",
         "copy-webpack-plugin": "^5.1.1",
