+++ src/components/Banner.jsx
@@ -5,8 +5,10 @@ const Banner = () => {
     <div className="govuk-width-container app-width-container--wide">
       <div className="govuk-phase-banner">
         <p className="govuk-phase-banner__content">
-          <strong className="govuk-tag govuk-phase-banner__content__tag">alpha</strong>
-          <span className="govuk-phase-banner__text">
+          <strong className="govuk-tag govuk-pahase-banner__content__tag">
+            alpha
+          </strong>
+          <span className="govuk-phase-banner__text" data-testid="banner-text">
             This is a new service – your
             {' '}
             <a className="govuk-link" href="#">feedback</a>
